requirements.txt
attrs==23.2.0
certifi==2024.6.2
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
PySocks==1.7.1
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
Faker==25.9.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
python-dateutil==2.9.0.post0
six==1.16.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
